Updated according to suggestions
